+++ .gitignore
@@ -17,3 +17,5 @@
 /.web-server-pid
 ###< symfony/web-server-bundle ###
 /.idea
+
+node_modules
