+++ public/assets/js/profissionais/busca-profissionais.js
@@ -1,11 +1,1 @@
-export class BuscaProfissionais {
-    constructor() {
-        this._url = '/api/professional/list';
-    }
-
-    busca() {
-        return fetch(this._url)
-            .then(res => res.json())
-            .then(profissionais => profissionais.filter(profissional => profissional.nome !== null));
-    }
-}
+export class BuscaProfissionais{constructor(){this._url="/api/professional/list"}busca(){return fetch(this._url).then(s=>s.json()).then(s=>s.filter(s=>null!==s.nome))}}
