+++ public/assets/js/profissionais/scripts.js
@@ -1,13 +1,1 @@
-import {BuscaProfissionais} from "./busca-profissionais.js";
-import {ViewProfissionais} from "./view-profissionais.js";
-
-const busca = new BuscaProfissionais();
-const view = new ViewProfissionais(document.getElementById('cards-profissionais'));
-document.querySelector('#especialidade').addEventListener('change', () => {
-    view.showLoader();
-    busca.busca()
-        .then(profisisonais => {
-            view.render(profisisonais);
-        })
-        .catch(err => console.log(err));
-});
+import{BuscaProfissionais}from"./busca-profissionais.js";import{ViewProfissionais}from"./view-profissionais.js";const busca=new BuscaProfissionais,view=new ViewProfissionais(document.getElementById("cards-profissionais"));document.querySelector("#especialidade").addEventListener("change",()=>{view.showLoader(),busca.busca().then(s=>{view.render(s)}).catch(s=>console.log(s))});
