+++ public/assets/js/solicitar-agendamento/buscador-origens.js
@@ -1,16 +1,1 @@
-export class BuscadorDeOrigens {
-    constructor(el) {
-        this._selectElement = el;
-    }
-
-    busca() {
-        fetch('/api/patient/list-sources')
-            .then(res => res.json())
-            .then(sources => {
-                this._selectElement.innerHTML = '<option value="">Selecione</option>';
-                sources
-                    .sort((a, b) => a.nome_origem < b.nome_origem ? -1 : 1)
-                    .forEach(source => this._selectElement.innerHTML += `<option value="${source.origem_id}">${source.nome_origem}</option>`);
-            });
-    }
-}
+export class BuscadorDeOrigens{constructor(e){this._selectElement=e}busca(){fetch("/api/patient/list-sources").then(e=>e.json()).then(e=>{this._selectElement.innerHTML='<option value="">Selecione</option>',e.sort((e,o)=>e.nome_origem<o.nome_origem?-1:1).forEach(e=>this._selectElement.innerHTML+=`<option value="${e.origem_id}">${e.nome_origem}</option>`)})}}
